Refactor and fix bad patterns
